Python2 drop (#109)
* remove python2 tox
* remove python2 dependencies and bump deps version
* removed the python2 switches in setup.py
* removed the python2 tox
* added the python3.8 toxenv
* fix circuit validation for h5py>=3.0.0
* remove all deprecates

+++ bluepysnap/edges.py
@@ -217,7 +217,7 @@ class EdgePopulation(object):
                     result[p] = self._get_property(p, selection)
         else:
             if len(edge_ids) == 0:
-                result = pd.Series(name=properties)
+                result = pd.Series(name=properties, dtype=np.float64)
             else:
                 result = pd.Series(
                     self._get_property(properties, selection),
